Refactor : 기존 repository 폴더 model + repository 로 분기
