+++ src/main/kotlin/com/everywaffle/team3server/user/repository/UserRepository.kt
@@ -1,5 +1,6 @@
 package com.everywaffle.team3server.user.repository
 
+import com.everywaffle.team3server.user.model.UserEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository : JpaRepository<UserEntity, Long>
